requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
PyJWT==1.7.1
PySocks==1.6.8
python-dotenv==0.10.1
python-http-client==3.1.0
pytz==2018.7
requests==2.21.0
sendgrid==5.6.0
six==1.12.0
twilio==6.22.0
urllib3==1.24.1
